+++ DomoticzAPI/__init__.py
@@ -16,3 +16,4 @@ from .setting import *
 from .settings import Settings
 from .translation import *
 from .devicetimer import *
+from .basetimer import *
